+++ src/ebic/auth/micro.py
@@ -41,8 +41,6 @@ class Permissions(GenericPermissions):
             headers={"Authorization": f"Bearer {token}"},
         )
 
-        print(response.status_code)
-
         if response.status_code != 200:
             raise HTTPException(
                 status_code=response.status_code, detail=response.json().get("detail")